footerのコンポネント化、状態管理をPAGE.JSXに一元管理でコードの整理
